LayoutTests/js/dfg-cos-constant-expected.txt

Tests that Math.cos() on a constant works in the DFG.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo() is 1 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
